+++ jgrapht-core/src/main/java/org/jgrapht/alg/cycle/UndirectedCycleBase.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2013-2017, by Nikolay Ognyanov and Contributors.
+ * (C) Copyright 2013-2018, by Nikolay Ognyanov and Contributors.
  *
  * JGraphT : a free Java graph-theory library
  *
@@ -20,6 +20,7 @@ package org.jgrapht.alg.cycle;
 import java.util.*;
 
 import org.jgrapht.*;
+import org.jgrapht.alg.interfaces.CycleBasisAlgorithm;
 
 /**
  * A common interface for classes implementing algorithms for finding a cycle base of an undirected
@@ -29,7 +30,9 @@ import org.jgrapht.*;
  * @param <E> the edge type.
  *
  * @author Nikolay Ognyanov
+ * @deprecated in favor of {@link CycleBasisAlgorithm}
  */
+@Deprecated
 public interface UndirectedCycleBase<V, E>
 {
     /**