+++ jgrapht-core/src/main/java/org/jgrapht/alg/shortestpath/DijkstraClosestFirstIterator.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2003-2017, by John V Sichi, Dimitrios Michail and Contributors.
+ * (C) Copyright 2003-2018, by John V Sichi, Dimitrios Michail and Contributors.
  *
  * JGraphT : a free Java graph-theory library
  *