+++ jgrapht-core/src/main/java/org/jgrapht/graph/ClassBasedEdgeFactory.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2003-2017, by Barak Naveh and Contributors.
+ * (C) Copyright 2003-2018, by Barak Naveh and Contributors.
  *
  * JGraphT : a free Java graph-theory library
  *
@@ -54,7 +54,7 @@ public class ClassBasedEdgeFactory<V, E>
     public E createEdge(V source, V target)
     {
         try {
-            return edgeClass.newInstance();
+            return edgeClass.getDeclaredConstructor().newInstance();
         } catch (Exception ex) {
             throw new RuntimeException("Edge factory failed", ex);
         }