+++ jgrapht-core/src/test/java/org/jgrapht/alg/shortestpath/KSPPathValidatorTest.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2016-2017, by Assaf Mizrachi and Contributors.
+ * (C) Copyright 2016-2018, by Assaf Mizrachi and Contributors.
  *
  * JGraphT : a free Java graph-theory library
  *
@@ -46,7 +46,7 @@ public class KSPPathValidatorTest
         SimpleGraph<String, DefaultEdge> clique = buildCliqueGraph(size);
         for (int i = 0; i < size; i++) {
             KShortestPaths<String, DefaultEdge> ksp = new KShortestPaths<String, DefaultEdge>(
-                clique, 1, Integer.MAX_VALUE, (prevPathElement, edge) -> false);
+                clique, 1, Integer.MAX_VALUE, (partialPath, edge) -> false);
 
             for (int j = 0; j < size; j++) {
                 if (j == i) {
@@ -68,7 +68,7 @@ public class KSPPathValidatorTest
         SimpleGraph<String, DefaultEdge> clique = buildCliqueGraph(size);
         for (int i = 0; i < size; i++) {
             KShortestPaths<String, DefaultEdge> ksp = new KShortestPaths<String, DefaultEdge>(
-                clique, 30, Integer.MAX_VALUE, (prevPathElement, edge) -> true);
+                clique, 30, Integer.MAX_VALUE, (partialPath, edge) -> true);
 
             for (int j = 0; j < size; j++) {
                 if (j == i) {
@@ -91,13 +91,13 @@ public class KSPPathValidatorTest
         SimpleGraph<Integer, DefaultEdge> ring = buildRingGraph(size);
         for (int i = 0; i < size; i++) {
             KShortestPaths<Integer, DefaultEdge> ksp = new KShortestPaths<Integer, DefaultEdge>(
-                ring, 2, Integer.MAX_VALUE, (prevPathElement, edge) -> {
-                    if (prevPathElement == null) {
+                ring, 2, Integer.MAX_VALUE, (partialPath, edge) -> {
+                    if (partialPath == null) {
                         return true;
                     }
                     return Math.abs(
-                        prevPathElement.getVertex() - Graphs
-                            .getOppositeVertex(ring, edge, prevPathElement.getVertex())) == 1;
+                        partialPath.getEndVertex() - Graphs
+                            .getOppositeVertex(ring, edge, partialPath.getEndVertex())) == 1;
                 });
 
             for (int j = 0; j < size; j++) {
@@ -122,7 +122,7 @@ public class KSPPathValidatorTest
         SimpleGraph<Integer, DefaultEdge> graph = buildGraphForTestDisconnected(cliqueSize);
         for (int i = 0; i < graph.vertexSet().size(); i++) {
             KShortestPaths<Integer, DefaultEdge> ksp = new KShortestPaths<Integer, DefaultEdge>(
-                graph, 100, Integer.MAX_VALUE, (prevPathElement, edge) -> {
+                graph, 100, Integer.MAX_VALUE, (partialPath, edge) -> {
                     // accept all requests but the one to pass through the edge connecting
                     // the two cliques.
                     DefaultEdge connectingEdge = graph.getEdge(cliqueSize - 1, cliqueSize);
@@ -146,6 +146,59 @@ public class KSPPathValidatorTest
 
             }
         }
+    }
+    
+    /**
+     * Testing that the provided GraphPath and new edge are generated correctly.
+     * On a directed line graph, the path at step i is expected to include all
+     * vertices [0..i-1] and edges {(0, 1), (1, 2), ... (i-1, i) and where
+     * new edge is (i, i+1). 
+     * v
+     */
+    public void testGraphPath()
+    {
+        SimpleDirectedGraph<Integer, DefaultEdge> line = buildLineGraph(10);
+        KShortestPaths<Integer, DefaultEdge> ksp = new KShortestPaths<Integer, DefaultEdge>(line, 
+            Integer.MAX_VALUE, new PathValidator<Integer, DefaultEdge>()
+        {
+
+            int index = 0;
+
+            @Override
+            public boolean isValidPath(
+                GraphPath<Integer, DefaultEdge> partialPath, DefaultEdge edge)
+            {
+                assertNotNull(edge);
+                assertEquals(line.getEdgeSource(edge), index, index + 1);
+
+                List<Integer> expectedVertices = new ArrayList<>();
+
+                for (int i = 0; i < index + 1; i++) {
+                    expectedVertices.add(i);
+                }
+                
+                List<DefaultEdge> expectedEdges = new ArrayList<>();
+                for (int i = 0; i < index; i++) {
+                    expectedEdges.add(line.getEdge(i, i + 1));
+                }
+                
+                assertNotNull(partialPath);
+                assertEquals(index, partialPath.getEdgeList().size());
+                assertEquals(expectedEdges, partialPath.getEdgeList());
+                assertEquals(index, partialPath.getEndVertex().intValue());
+                assertEquals(line, partialPath.getGraph());
+                assertEquals(index, partialPath.getLength());
+                assertEquals(0, partialPath.getStartVertex().intValue());
+                assertEquals(index + 1, partialPath.getVertexList().size());
+                assertEquals(expectedVertices, partialPath.getVertexList());
+                assertEquals((double) index, partialPath.getWeight());
+
+                index++;
+                return true;
+            }
+        });
+
+        ksp.getPaths(0, 9);
     }
 
     private SimpleGraph<String, DefaultEdge> buildCliqueGraph(int size)
@@ -198,4 +251,13 @@ public class KSPPathValidatorTest
         graphGenerator.generateGraph(clique, new IntegerVertexFactory(), null);
         return clique;
     }
+    
+    private SimpleDirectedGraph<Integer, DefaultEdge> buildLineGraph(int size)
+    {
+        SimpleDirectedGraph<Integer, DefaultEdge> line = new SimpleDirectedGraph<>(DefaultEdge.class);
+        LinearGraphGenerator<Integer, DefaultEdge> graphGenerator = new LinearGraphGenerator<>(size);
+        graphGenerator.generateGraph(line, new IntegerVertexFactory(), null);
+        return line;
+    }
+
 }